requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
PyPDF2>=3.0.1
scikit-learn>=1.4.0
requests>=2.32.0
python-dotenv>=1.0.1
pytest>=8.0.0
